refactor: split evidence.rs god-file into modules
