+++ src/core/runner/evidence/tests/mod.rs
@@ -1,0 +1,47 @@
+use crate::core::api_jobs::{Job, JobStatus};
+use crate::core::runner::{Runner, RunnerKind};
+use crate::core::server::{RunnerPolicy, RunnerSettings};
+use uuid::Uuid;
+
+mod artifact;
+mod download;
+mod mirror;
+
+pub(crate) fn ssh_runner() -> Runner {
+    Runner {
+        id: "lab".to_string(),
+        kind: RunnerKind::Ssh,
+        server_id: Some("srv".to_string()),
+        workspace_root: Some("/srv/homeboy".to_string()),
+        settings: RunnerSettings {
+            daemon: true,
+            ..Default::default()
+        },
+        env: Default::default(),
+        secret_env: Default::default(),
+        resources: Default::default(),
+        policy: RunnerPolicy::default(),
+    }
+}
+
+pub(crate) fn succeeded_job(id: Uuid) -> Job {
+    Job {
+        id,
+        operation: "exec".to_string(),
+        status: JobStatus::Succeeded,
+        created_at_ms: 1_700_000_000_000,
+        updated_at_ms: 1_700_000_001_000,
+        started_at_ms: Some(1_700_000_000_000),
+        finished_at_ms: Some(1_700_000_001_000),
+        event_count: 0,
+        source_snapshot: None,
+        stale_reason: None,
+        target_runner_id: None,
+        target_project_id: None,
+        claim_id: None,
+        claimed_by_runner_id: None,
+        claimed_at_ms: None,
+        claim_expires_at_ms: None,
+        artifacts: Vec::new(),
+    }
+}
